refactor(eve): split Slack input authorization
